+++ package.json
@@ -12,7 +12,7 @@
     "cors": "2.8.5",
     "cross-env": "7.0.3",
     "dotenv": "^16.0.3",
-    "express": "4.17.1",
+    "express": "^4.18.2",
     "joi": "^17.8.3",
     "mongoose": "^6.10.0",
     "morgan": "1.10.0",
