+++ routes/api/contacts.js
@@ -1,7 +1,7 @@
-const express = require('express');
+const express = require("express")
 const router = express.Router();
 
-const ctrl = require("../../controllers/contacts");
+const {contacts: ctrl} = require("../../controllers");
 const {contactShema, updateFavorite} = require("../../shemas");
 const {contactsValidate, isValidId} = require("../../middlewars");
