+++ routes/api/index.js
@@ -1,0 +1,5 @@
+const contactsRouter = require("./contacts")
+
+module.exports = {
+	contactsRouter,
+}
